stream_libarchive: fix libarchive callback signature
libarchive uses a quite confusing ifdeffery mess for some of the types
used in callbacks. Currently, archive_read_set_seek_callback() causes a
warning at least on Windows due to mismatching return type. The header
file uses __LA_INT64_T as return type, so I think the user is intended
to use int64_t.

(The ssize_t return type for the read_cb seems correct, on the other
hand.)

+++ stream/stream_libarchive.c
@@ -31,7 +31,7 @@ static ssize_t read_cb(struct archive *arch, void *priv, const void **buffer)
     return MPMAX(res, 0);
 }
 
-static ssize_t seek_cb(struct archive *arch, void *priv,
+static int64_t seek_cb(struct archive *arch, void *priv,
                        int64_t offset, int whence)
 {
     struct mp_archive *mpa = priv;
